rename index to camera
