+++ tfjs-react-native/src/camera/camera.ts
