+++ tfjs-react-native/src/index.ts
@@ -21,4 +21,4 @@ export {asyncStorageIO} from './async_storage_io';
 export {bundleResourceIO} from './bundle_resource_io';
 export {decodeJpeg} from './decode_image';
 export {fetch} from './platform_react_native';
-export * from './camera';
+export * from './camera/camera';
